backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Grigory Malanchuk <grigorymalanchuk@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.6"
djangorestframework = "^3.15.1"
python-dotenv = "^1.0.1"
psycopg2 = "^2.9.9"
djangorestframework-simplejwt = "^5.3.1"
djoser = "^2.2.3"
drf-spectacular = "^0.27.2"
django-debug-toolbar = "^4.4.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
